+++ delta-kernel-engine/src/default/json.rs
